Cambios 22-03-2021

+++ DistribuidoraVendedores/DistribuidoraVendedores/Models/DetalleCompra_previo.cs
@@ -12,7 +12,7 @@ namespace DistribuidoraVendedores.Models
         public decimal precio_producto { get; set; }
         public decimal descuento { get; set; }
         public decimal sub_total { get; set; }
-        public decimal stock { get; set; }
+        public int stock { get; set; }
         public decimal stock_valorado { get; set; }
         public decimal promedio { get; set; }
     }
