+++ DistribuidoraVendedores/DistribuidoraVendedores/Models/DetalleVenta_previo.cs
@@ -13,7 +13,7 @@ namespace DistribuidoraVendedores.Models
         public decimal descuento { get; set; }
         public int envases { get; set; }
         public decimal sub_total { get; set; }
-        public decimal stock { get; set; }
+        public int stock { get; set; }
         public decimal stock_valorado { get; set; }
         public decimal promedio { get; set; }
     }
